Rename a bunch of DutyStation things to DutyLocation
