+++ src/components/DutyLocationSearchBox/DutyLocationSearchBox.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import DutyStationSearchBox from './DutyStationSearchBox';
+import DutyStationSearchBox from './DutyLocationSearchBox';
 
 const testAddress = {
   city: 'Glendale Luke AFB',
