+++ src/scenes/Review/EditProfile.jsx
@@ -12,7 +12,7 @@ import Alert from 'shared/Alert';
 import { SwaggerField } from 'shared/JsonSchemaForm/JsonSchemaField';
 import { validateAdditionalFields } from 'shared/JsonSchemaForm';
 import SaveCancelButtons from './SaveCancelButtons';
-import DutyStationSearchBox from 'components/DutyStationSearchBox/DutyStationSearchBox';
+import DutyStationSearchBox from 'components/DutyLocationSearchBox/DutyLocationSearchBox';
 import scrollToTop from 'shared/scrollToTop';
 import {
   selectCurrentMove,
