+++ src/types/index.js
@@ -1,6 +1,6 @@
 export { UserRoleShape, UserRolesShape, OfficeUserInfoShape, TransportationOfficeShape } from './user';
 export { AddressShape } from './address';
-export { DutyStationShape } from './dutyStation';
+export { DutyLocationShape } from './dutyLocation';
 export { DropdownArrayOf } from './form';
 export {
   DestinationDutyStationShape,
